requirements.txt
pytest
requests
jsonschema
pytest-xdist
allure-pytest
PyYAML
